[fix] 결제내역 상세보기

See merge request s10-fintech-finance-sub2/S10P22C203!200

+++ BackEnd/src/main/java/com/orange/fintech/payment/entity/Receipt.java
@@ -39,9 +39,9 @@ public class Receipt {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm:ss", timezone = "Asia/Seoul")
     private LocalTime transactionTime;
 
-    @NotNull private int totalPrice;
+    @NotNull private Long totalPrice;
 
-    @NotNull private int approvalAmount;
+    @NotNull private Long approvalAmount;
 
     private long authNumber;
